IS: Refactor queries to dialogmotekandidatVurderingRepository
