+++ src/main/kotlin/no/nav/syfo/application/DialogmotekandidatVurderingService.kt
@@ -1,16 +1,14 @@
 package no.nav.syfo.application
 
 import no.nav.syfo.api.exception.ConflictException
-import no.nav.syfo.domain.DialogmotekandidatEndring
-import no.nav.syfo.domain.IkkeAktuell
-import no.nav.syfo.domain.Personident
-import no.nav.syfo.domain.isLatestIkkeKandidat
+import no.nav.syfo.domain.*
 import no.nav.syfo.infrastructure.database.DatabaseInterface
 import no.nav.syfo.infrastructure.database.dialogmotekandidat.getDialogmotekandidatEndringListForPerson
 import no.nav.syfo.infrastructure.database.dialogmotekandidat.toDialogmotekandidatEndringList
 import no.nav.syfo.infrastructure.database.toIkkeAktuellList
+import no.nav.syfo.infrastructure.database.toUnntakList
 
-class IkkeAktuellService(
+class DialogmotekandidatVurderingService(
     private val database: DatabaseInterface,
     private val dialogmotekandidatService: DialogmotekandidatService,
     private val dialogmotekandidatVurderingRepository: IDialogmotekandidatVurderingRepository,
@@ -51,4 +49,39 @@ class IkkeAktuellService(
 
     suspend fun getIkkeAktuellList(personIdent: Personident): List<IkkeAktuell> =
         dialogmotekandidatVurderingRepository.getIkkeAktuellListForPerson(personIdent = personIdent).toIkkeAktuellList()
+
+    suspend fun createUnntak(
+        unntak: Unntak,
+        veilederToken: String,
+        callId: String,
+    ) {
+        val latestOppfolgingstilfelleArbeidstaker = oppfolgingstilfelleService.getLatestOppfolgingstilfelle(
+            arbeidstakerPersonIdent = unntak.personIdent,
+            veilederToken = veilederToken,
+            callId = callId,
+        )
+        database.connection.use { connection ->
+            val ikkeKandidat =
+                connection.getDialogmotekandidatEndringListForPerson(personIdent = unntak.personIdent)
+                    .toDialogmotekandidatEndringList()
+                    .isLatestIkkeKandidat()
+            if (ikkeKandidat) {
+                throw ConflictException("Failed to create Unntak: Person is not kandidat")
+            }
+            dialogmotekandidatVurderingRepository.createUnntak(connection = connection, unntak = unntak)
+            val newDialogmotekandidatEndring = DialogmotekandidatEndring.unntak(
+                personIdentNumber = unntak.personIdent,
+            )
+            dialogmotekandidatService.createDialogmotekandidatEndring(
+                connection = connection,
+                dialogmotekandidatEndring = newDialogmotekandidatEndring,
+                tilfelleStart = latestOppfolgingstilfelleArbeidstaker?.tilfelleStart,
+                unntak = unntak,
+            )
+            connection.commit()
+        }
+    }
+
+    suspend fun getUnntakList(personIdent: Personident): List<Unntak> =
+        dialogmotekandidatVurderingRepository.getUnntakList(personIdent = personIdent).toUnntakList()
 }
